requirements.txt
Appium-Python-Client==3.2.1
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
decorator==5.2.1
docutils==0.21.2
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
Jinja2==3.1.6
kitchen==1.2.6
MarkupSafe==3.0.2
outcome==1.3.0.post0
pathspec==0.12.1
platformdirs==4.3.6
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
robotframework==7.2.2
robotframework-appiumlibrary==2.1.0
robotframework-robocop==5.8.1
selenium==4.16.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tomli==2.2.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==1.26.20
wsproto==1.2.0
